update_extended_position_control

+++ scripts/velocity_control.py
@@ -6,9 +6,9 @@ from std_msgs.msg import String
 from sensor_msgs.msg import JointState
 from math import pi
 
-wheel_size_mm = 42
-max_rpm = 70
-limit_speed_rpm = lambda input_rpm : min(max( int(input_rpm)/60*2*pi , -max_rpm/60*2*pi), max_rpm/60*2*pi)   
+wheel_size_mm = 42.0
+max_rpm = 70.0
+limit_speed_rpm = lambda input_rpm : min(max( float(input_rpm)/60.0*2*pi , -max_rpm/60.0*2*pi), max_rpm/60.0*2*pi)   
 # speed =  1/.229  ==> 41rpm / 2/3 rps 
 
 speed = lambda input : 70/60*2*pi    
@@ -38,7 +38,7 @@ def talker():
 
     while not rospy.is_shutdown():
    
-        input_rad_p_s = limit_speed_rpm(input('input rpm(max is 70)'))
+        input_rad_p_s = limit_speed_rpm(float(input('input rpm(max is 70)')))
         rospy.loginfo("set speed {rpm:.2f} rpm, (approx) {mmps:.2f} mm/s".format(rpm=input_rad_p_s*60/(2*pi), mmps=(input_rad_p_s/(2*pi)*pi*wheel_size_mm) ))
